backend/pyproject.toml
[project]
name = "moana"
version = "0.1.0"
description = "AI-native early childhood education content generation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "python-multipart>=0.0.6",
    "alembic>=1.13.0",
    "tenacity>=8.2.0",
    "anthropic>=0.18.0",
    # Phase 6 - Auth & Security
    "PyJWT>=2.8.0",
    # Phase 6 - Image Processing
    "Pillow>=10.0.0",
    "qrcode>=7.4.0",
    # Phase 6 - OSS Storage (optional)
    "oss2>=2.18.0",
    # Phase 6 - Google Gemini
    "google-generativeai>=0.3.0",
    # Phase 7 - Google Imagen & Veo
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
